pyproject.toml
[tool.poetry]
name = "minimal-provider-agent-market"
version = "1.0.0"
description = "A streamlined Python service for the Agent Market platform"
authors = ["OpenHands <openhands@all-hands.dev>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
httpx = "0.27.2"
loguru = "^0.7.3"
pydantic = "2.10.3"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
openai = "1.57.0"
aider-chat = "^0.70.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
